Attempt to make libuv fork safe.
If threadpool is initialized try and make libuv fork safe. As a library
it's hard to predict what an application may do and a programmer not
realizing libuv may spawn threads may fork without an exec call. If this
occurs and threadpool is initialized the child will abort in shutdown as
the threads no longer exist and pthread_join fails.

This checkin doesn't work at the moment - the threadpool is now fork
safe, but on BSD kqueue is not and any loop will have an invalid kqueue
in the child.